Switched DTO coupling from an Interface to an Abstract class
