+++ src/main/java/model/AbstractDtoObject.java
@@ -1,0 +1,33 @@
+package model;
+
+public abstract class AbstractDtoObject {
+    private int id;
+    private String name;
+    public AbstractDtoObject() {}
+
+    public AbstractDtoObject(String name, int id) {
+        this.id = id;
+        this.name = name;
+    }
+
+    public int getId() {
+        return id;
+    }
+
+    public void setId(int id) {
+        this.id = id;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String toString() {
+        return name;
+    }
+}
