--- src/main/java/model/DtoObjectInterface.java
@@ -1,5 +1,0 @@
-package model;
-
-public interface DtoObjectInterface {
-
-}
