+++ src/main/java/model/ModelDto.java
@@ -1,40 +1,14 @@
 package model;
 
-import com.google.gson.Gson;
-import com.google.gson.reflect.TypeToken;
-
-import java.util.List;
-
-public class ModelDto implements DtoObjectInterface {
-
-    private int id;
-    private String name;
+public class ModelDto extends AbstractDtoObject {
     private CategoryDto category;
 
     public ModelDto() {}
 
     public ModelDto(CategoryDto category, String name, int id) {
-        this.id = id;
+        super(name, id);
         this.category = category;
-        this.name = name;
     }
-
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
 
     public CategoryDto getCategory() {
         return category;
@@ -42,10 +16,5 @@ public class ModelDto implements DtoObjectInterface {
 
     public void setCategory(CategoryDto category) {
         this.category = category;
-    }
-
-    @Override
-    public String toString() {
-        return name;
     }
 }
